mikanani/requirements.txt
requests
loguru
beautifulsoup4
python-qbittorrent~=0.4.3
dotenv==1.0.0
pyyaml==6.0.1
xmltodict==0.13.0